requirements.txt
Click>=6.0
python-dateutil==2.8.0
inflection==0.3.1
PyJWT==1.7.1
tabulate==0.8.3
jsonschema==3.0.1
